Adding unique key
An error/warning pops up the browser console: 'Each child in a list should have a unique "key" prop.'

To fix this we add a key to each list item using their index in the array.

If we had an array of objects with id's we could have used those.